delete not found

+++ src/main/java/org/nruharish/springmvc/services/BeerServiceJPA.java
@@ -65,8 +65,12 @@ public class BeerServiceJPA implements BeerService{
     }
 
     @Override
-    public void deleteBeerById(UUID id) {
-        beerRepository.deleteById(id);
+    public Boolean deleteBeerById(UUID id) {
+        if(beerRepository.existsById(id)) {
+            beerRepository.deleteById(id);
+            return true;
+        }
+        return false;
 
     }
